refactor(blorktools): rig hoist again
